+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `--args <file.yml>` argument insertion
 
+### Fixed
+
+- `style_text` now adapts Tr (text rendering mode) better to user parameters
+
 ## [0.24.0] - 2026-07-07
 
 ### Added